% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_element_rules}
\alias{apply_element_rules}
\title{Apply elemental composition checks (carbon count)}
\usage{
apply_element_rules(curdata, current_conf)
}
\description{
Apply elemental composition checks (carbon count)
}
