% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_final_adjustments}
\alias{apply_final_adjustments}
\title{Apply final adduct-based adjustments}
\usage{
apply_final_adjustments(curdata, cur_adducts, filter.by, current_conf)
}
\description{
Apply final adduct-based adjustments
}
