% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_module_rules}
\alias{apply_module_rules}
\title{Apply rules based on multiple RT modules}
\usage{
apply_module_rules(curdata, cur_adducts, filter.by, current_conf)
}
\description{
Apply rules based on multiple RT modules
}
