% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_multimer_rules}
\alias{apply_multimer_rules}
\title{Apply multimer and charge state abundance checks}
\usage{
apply_multimer_rules(
  curdata,
  cur_adducts,
  adduct_table_local,
  current_conf,
  multimer_abundance_check
)
}
\description{
Apply multimer and charge state abundance checks
}
