% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_single_row_rules}
\alias{apply_single_row_rules}
\title{Apply rules for single row data}
\usage{
apply_single_row_rules(
  curdata,
  cur_adducts,
  adduct_weights,
  filter.by,
  current_conf
)
}
\description{
Apply rules for single row data
}
