% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{apply_unique_adduct_rules}
\alias{apply_unique_adduct_rules}
\title{Check confidence based on unique adduct count}
\usage{
apply_unique_adduct_rules(curdata, cur_adducts, filter.by, current_conf)
}
\description{
Check confidence based on unique adduct count
}
