% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_boosted_compounds_table}
\alias{as_boosted_compounds_table}
\title{Validate boosted compounds table}
\usage{
as_boosted_compounds_table(data, match_by = c("mz", "rt"))
}
\arguments{
\item{data}{Boosted compounds with columns: compound_id (required), mz (optional), rt (optional)}

\item{match_by}{Which columns to use for matching: c("mz"), c("rt"), or c("mz", "rt")}
}
\value{
Validated table with ID, mz, time columns for boost_confidence_of_IDs()
}
\description{
Validate boosted compounds table
}
