% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{assign_conf}
\alias{assign_conf}
\title{Assign confidence based on standard criteria}
\usage{
assign_conf(curdata, filter.by, delta_rt, max_diff_rt, current_conf)
}
\description{
Assign confidence based on standard criteria
}
