% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{boost_confidence_of_IDs}
\alias{boost_confidence_of_IDs}
\title{Boost confidence of specified compound IDs to level 4}
\usage{
boost_confidence_of_IDs(
  chemscoremat_with_confidence,
  boostIDs,
  max.mz.diff,
  max.rt.diff
)
}
\arguments{
\item{chemscoremat_with_confidence}{Annotation results}

\item{boostIDs}{Data frame with ID column (and optional mz, time)}

\item{max.mz.diff}{Mass tolerance for matching}

\item{max.rt.diff}{RT tolerance for matching}
}
\value{
Updated data frame with boosted confidences
}
\description{
Boost confidence of specified compound IDs to level 4
}
