% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{check_minimum_score}
\alias{check_minimum_score}
\title{Initial confidence check based on score threshold
Returns NULL if should continue, or a result data frame if should return early}
\usage{
check_minimum_score(curdata)
}
\description{
Initial confidence check based on score threshold
Returns NULL if should continue, or a result data frame if should return early
}
