% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{compute_cluster_assignments}
\alias{compute_cluster_assignments}
\title{Compute the assignment of cluster ID based on nn-algorithm.}
\usage{
compute_cluster_assignments(clusters, data)
}
\arguments{
\item{clusters}{Position of cluster centers.}

\item{data}{Data to assign to clusters}
}
\value{
Indices of closest cluster for each query point computed using \code{\link[RANN:nn2]{RANN::nn2()}}.
}
\description{
Compute the assignment of cluster ID based on nn-algorithm.
}
