% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{compute_cluster_positions}
\alias{compute_cluster_positions}
\title{Compute cluster positions in data}
\usage{
compute_cluster_positions(data, width = 1, kernel = "gaussian", show = FALSE)
}
\arguments{
\item{data}{Data for which to compute the clustering.}

\item{width}{Bandwidth of the clustering.}

\item{kernel}{Kernel to use. See \code{\link[stats:density]{stats::density()}} for details.}

\item{show}{If TRUE, plot the resulting clustering.
Scan rate should be consulted to optimize clustering. \link{0.5;2.0}}
}
\value{
Positions of dense clusters.
}
\description{
Compute cluster positions in data
}
