% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{compute_delta_ppm}
\alias{compute_delta_ppm}
\title{Compute delta ppm between observed and theoretical m/z (vectorized)}
\usage{
compute_delta_ppm(chemscoremat_with_confidence)
}
\arguments{
\item{chemscoremat_with_confidence}{Data frame with mz and theoretical.mz columns}
}
\value{
Data frame with delta_ppm column inserted after column 8
}
\description{
Compute delta ppm between observed and theoretical m/z (vectorized)
}
