% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{compute_dominant_modules}
\alias{compute_dominant_modules}
\title{For each compound computes most abundant modules}
\usage{
compute_dominant_modules(annotations)
}
\arguments{
\item{annotations}{data frame with annotations}
}
\value{
tibble with \emph{compound}, \emph{dominant_module} columns
}
\description{
For each annotated compound computes a set (ie. more then one) of most
abundant (dominant) modules. The abundance of modules is computed from
distinct adduct-compound-peak tripplets.
}
