% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotopes.R
\name{compute_isotopic_pattern}
\alias{compute_isotopic_pattern}
\title{Compute isotopic pattern of a given molecule.}
\usage{
compute_isotopic_pattern(formula, minAbund = 0.001)
}
\arguments{
\item{formula}{A string containing molecular or empirical formula of a compound.}

\item{minAbund}{A cutoff abundance ratio: only isotopes with abundance ratio above this value are returned.}
}
\value{
A dataframe containing isotopic pattern of a given molecule. The pattern is represented by three columns:
mass, mass number difference, and abundance. Abundance of the isotopes is normalized with the most abundant isotope
having a unit value and abundances of the other isotopes are represented as a share of the most abundant isotope.
Mass number difference is 0 for the most abundant isotope.
}
\description{
Compute isotopic pattern of a given molecule.
}
