% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{compute_num_kernel_points}
\alias{compute_num_kernel_points}
\title{Compute number of equally spaced kernel points for density estimation.}
\usage{
compute_num_kernel_points(data)
}
\arguments{
\item{data}{Data for which to estimate the number of kernel points.}
}
\value{
Nearest power of 2 to length(data), but at least 2048.
}
\description{
Compute number of equally spaced kernel points for density estimation.
}
