% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{compute_pathway_scores}
\alias{compute_pathway_scores}
\title{For each pathway compute a score}
\usage{
compute_pathway_scores(pathways)
}
\arguments{
\item{pathways}{}
}
\value{
pathway table with a new pathway_score column
}
\description{
For each pathway compute a score
}
