% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{compute_pathways}
\alias{compute_pathways}
\title{Based on pathway data adjust the scores.}
\usage{
compute_pathways(
  annotations,
  pathways,
  adduct_weights,
  exluded_pathways = NULL,
  exluded_pathway_compounds = NULL,
  score_threshold = 0.1
)
}
\arguments{
\item{annotations}{Table with feature annotations}

\item{pathways}{Database of pathways to check for matching annotations}

\item{adduct_weights}{Adduct weights table for filtering valid adducts}

\item{exluded_pathways}{Pathways that are excluded from the matching}

\item{exluded_pathway_compounds}{Compounds which should be excluded from pathway analysis}

\item{score_threshold}{Score threshold to use to filter annotations}
}
\value{
annotations
}
\description{
Based on pathway data adjust the scores.
}
