% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_peak_modules.R
\name{compute_peak_correlations}
\alias{compute_peak_correlations}
\title{Compute correlation matrix for intensity table.}
\usage{
compute_peak_correlations(peak_intensity_matrix, correlation_method = "p")
}
\arguments{
\item{peak_intensity_matrix}{Intensity matrix.}

\item{correlation_method}{Correlation method, default "p" -> pearson}
}
\value{
Peak correlation matrix.
}
\description{
Compute correlation matrix for intensity table.
}
