% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{compute_peak_indices}
\alias{compute_peak_indices}
\title{Compute the positions of peak indices for given kernel density estimate.}
\usage{
compute_peak_indices(density)
}
\arguments{
\item{density}{y attribute from kernel density estimates \code{\link[stats:density]{stats::density()}}.}
}
\value{
Indices at which kernel density estimate position is a cluster.
}
\description{
Compute the positions of peak indices for given kernel density estimate.
}
