% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_peak_modules.R
\name{compute_peak_modules}
\alias{compute_peak_modules}
\title{Performs clustering analysis on top of a peak intensity matrix}
\usage{
compute_peak_modules(
  peak_intensity_matrix,
  peak_correlation_matrix,
  correlation_threshold,
  deep_split,
  min_cluster_size,
  network_type
)
}
\arguments{
\item{peak_intensity_matrix}{matrix of peak intensities. Each column
represents intensities of one peak in different samples, each row represent
intensitiess of different peaks in one sample}

\item{peak_correlation_matrix}{correlation matrix of peak intensities}

\item{correlation_threshold}{see the \emph{cutHeight} parameter of
\link{WGCNA::mergeCloseModules}}

\item{deep_split}{see \link{WGCNA::mergeCloseModules}}

\item{min_cluster_size}{see \link{WGCNA::blockwiseModules},
\link{WGCNA::mergeCloseModules}, and \link{dynamicTreeCut::cutreeDynamic}}

\item{network_type}{see \link{WGCNA::adjacency}}
}
\value{
data frame with 3 columns: \emph{peak}, \emph{mean_intensity}, and
\emph{module}
}
\description{
Performs clustering analysis on top of a peak intensity matrix
}
