% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{compute_rt_modules}
\alias{compute_rt_modules}
\title{Compute RT modules based on intensity modules and chromatographic peak width.}
\usage{
compute_rt_modules(peak_table, peak_width = 1, show = FALSE)
}
\arguments{
\item{peak_table}{Feature table with columns \link{'rt', 'module','peak'}.}

\item{peak_width}{Estimated chromatographic peak width.}

\item{show}{If TRUE, plot the resulting clustering.}
}
\value{
data frame with 4 columns: \emph{peak}, \emph{mean_intensity},
\emph{module} and \emph{rt_cluster}
}
\description{
Compute RT modules based on intensity modules and chromatographic peak width.
}
