% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute a confidence score for each annotation.}
\usage{
compute_scores(annotation, adduct_weights)
}
\arguments{
\item{annotation}{A table with annotated peaks.}

\item{adduct_weights}{A weight-by-adduct table.}
}
\description{
Compute a confidence score for each annotation.
}
