% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_utils.R
\name{construct_adduct_formula}
\alias{construct_adduct_formula}
\title{Recombine adduct formula and mass number difference.}
\usage{
construct_adduct_formula(formula, mass_number_difference)
}
\arguments{
\item{formula}{Adduct formula, e.g 'M+H'}

\item{mass_number_difference}{Difference from monoisotopic mass.}
}
\value{
Combined formula, e.g 'M+H_\link{+1}'.
}
\description{
Recombine adduct formula and mass number difference.
}
