% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{count_filter_matches}
\alias{count_filter_matches}
\title{Count how many adducts match the filter list}
\usage{
count_filter_matches(adducts, filter.by)
}
\description{
Count how many adducts match the filter list
}
