% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{create_cluster_table}
\alias{create_cluster_table}
\title{Create frequency table of clusters, keeping only the most frequent}
\usage{
create_cluster_table(curdata)
}
\description{
Create frequency table of clusters, keeping only the most frequent
}
