% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotopes.R
\name{detect_isotopic_peaks}
\alias{detect_isotopic_peaks}
\title{Find all possible isotopes of a given annotated peak by filtering peaks based on several criteria
(see \link{filter_isotopes}) and matching the intensities with natural abundances of the isotopes.}
\usage{
detect_isotopic_peaks(
  ...,
  intensity_deviation_tolerance,
  peaks,
  mass_defect_tolerance,
  rt_tolerance,
  isotope_mass_tolerance_ppm = NULL
)
}
\arguments{
\item{...}{A single annotated feature passed as a list of columns. See \code{\link[purrr:pmap]{purrr::pmap()}} for more details.}

\item{intensity_deviation_tolerance}{A numeric threshold by which an intensity ratio of two isotopic peaks may differ
from their actual abundance ratio.}

\item{peaks}{A peak table containing a peak identifier (unique number), mean intensity, module, and rt cluster
of each identified peak.}

\item{mass_defect_tolerance}{A number. Maximum difference in mass defect between two peaks of the same compound.}

\item{rt_tolerance}{A number. Maximum rt difference for two peaks of the same substance.}

\item{isotope_mass_tolerance_ppm}{Maximum allowed mass error in ppm between observed and theoretical isotope m/z.
If NULL, no ppm filter is applied.}
}
\value{
A table with peaks that have been identified as isotopes of a given molecule from the annotation table.
}
\description{
Find all possible isotopes of a given annotated peak by filtering peaks based on several criteria
(see \link{filter_isotopes}) and matching the intensities with natural abundances of the isotopes.
}
