% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_permutation.R
\name{detect_isotopic_peaks_cached}
\alias{detect_isotopic_peaks_cached}
\title{Detect isotopic peaks using precomputed pattern cache}
\usage{
detect_isotopic_peaks_cached(
  ...,
  intensity_deviation_tolerance,
  peaks,
  mass_defect_tolerance,
  rt_tolerance,
  isotope_mass_tolerance_ppm = NULL,
  pattern_cache
)
}
\arguments{
\item{...}{A single annotated feature passed as a list of columns}

\item{intensity_deviation_tolerance}{Intensity deviation tolerance}

\item{peaks}{Peak table}

\item{mass_defect_tolerance}{Mass defect tolerance}

\item{rt_tolerance}{RT tolerance}

\item{isotope_mass_tolerance_ppm}{Mass tolerance in ppm}

\item{pattern_cache}{Named list of precomputed isotope patterns}
}
\value{
Table of identified isotopes
}
\description{
Same as detect_isotopic_peaks but uses a precomputed pattern cache
instead of computing patterns on-the-fly.
}
\keyword{internal}
