% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{estimate_kernel_density}
\alias{estimate_kernel_density}
\title{Compute kernel density estimate.}
\usage{
estimate_kernel_density(data, width = 1, kernel = "gaussian")
}
\arguments{
\item{data}{Data for which to estimate the kernel density.}

\item{width}{Kernel width to use. Default 1. \link{0.5;2}}

\item{kernel}{Kernel to use. See stats::density for details.
Recommended \link{"gaussian","biweight","cosine","epanechnikov"}.}
}
\value{
Kernel density estimate \code{\link[stats:density]{stats::density()}}.
}
\description{
Compute kernel density estimate.
}
