% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{filter_clusters}
\alias{filter_clusters}
\title{Filter data to keep only rows matching the most frequent cluster}
\usage{
filter_clusters(curdata, table_names)
}
\description{
Filter data to keep only rows matching the most frequent cluster
}
