% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotopes.R
\name{filter_isotopes}
\alias{filter_isotopes}
\title{Match isotopes from peak table to annotated peaks based on rt cluster, rt difference, and mass defect.}
\usage{
filter_isotopes(
  query,
  intensity_deviation_tolerance,
  peaks,
  mass_defect_tolerance,
  rt_tolerance
)
}
\arguments{
\item{query}{A single annotated peak.}

\item{intensity_deviation_tolerance}{A numeric threshold by which an intensity ratio of two isotopic peaks may differ
from their actual abundance ratio.}

\item{peaks}{A peak table containing a peak identifier (unique number), mean intensity, module, and rt cluster
of each identified peak.}

\item{mass_defect_tolerance}{A number. Maximum difference in mass defect between two peaks of the same compound.}

\item{rt_tolerance}{A number. Maximum rt difference for two peaks of the same substance.}
}
\value{
A table of identified isotopes with an extra column: \code{mass_number_difference}.
The extra column represents a nominal mass (mass number) difference between a given isotope and the most abundant
isotope of the same molecule.
}
\description{
Match isotopes from peak table to annotated peaks based on rt cluster, rt difference, and mass defect.
}
