% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotopes.R
\name{get_monoisotopic_names}
\alias{get_monoisotopic_names}
\title{Identify monoisotopic element names from enviPat column headers.}
\usage{
get_monoisotopic_names(element_cols)
}
\arguments{
\item{element_cols}{Character vector of enviPat element column names.}
}
\value{
Character vector of monoisotopic element names present in the input.
}
\description{
enviPat returns columns like "12C", "13C", "1H", "2H", "14N", "15N", etc.
This function returns the names of the most abundant (monoisotopic) isotope
for each element, so that heavy isotope substitutions can be identified.
}
\keyword{internal}
