% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peak_intensity_matrix}
\alias{get_peak_intensity_matrix}
\title{Get only intensity part of peak table with peaks as columns and samples on rows.}
\usage{
get_peak_intensity_matrix(peak_table)
}
\arguments{
\item{peak_table}{Peak table from which to extract the intensities.}
}
\value{
Intensities with samples on rows and peaks on columns.
}
\description{
Get only intensity part of peak table with peaks as columns and samples on rows.
}
