% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{is_filter_empty}
\alias{is_filter_empty}
\title{Check if filter.by is effectively empty (NULL or single NA)}
\usage{
is_filter_empty(filter.by)
}
\description{
Check if filter.by is effectively empty (NULL or single NA)
}
