% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isotopes.R
\name{match_isotopes_by_intensity}
\alias{match_isotopes_by_intensity}
\title{Match prefiltered isotopes by comparing their relative intensity to their relative natural abundance.
The relative values are computed as a ratio of an isotopic peak intensity to the intensity of its monoisotopic peak,
respectively as a ratio of a natural abundance of the isotope to the abundance of the most abundant isotope.}
\usage{
match_isotopes_by_intensity(
  query,
  isotopes,
  pattern,
  intensity_deviation_tolerance,
  isotope_mass_tolerance_ppm = NULL
)
}
\arguments{
\item{query}{A single annotated peak.}

\item{isotopes}{Isotope candidates for the annotated molecule. The candidates are identified in \link{compute_isotopes}.}

\item{pattern}{Isotopic pattern of a given molecule as produced by \link{compute_isotopic_pattern}.}

\item{intensity_deviation_tolerance}{A numeric threshold by which an intensity ratio of two isotopic peaks may differ
from their actual abundance ratio.}

\item{isotope_mass_tolerance_ppm}{Maximum allowed mass error in ppm between observed and theoretical isotope m/z.
If NULL, no ppm filter is applied.}
}
\value{
A table of matched isotopes.
}
\description{
Match prefiltered isotopes by comparing their relative intensity to their relative natural abundance.
The relative values are computed as a ratio of an isotopic peak intensity to the intensity of its monoisotopic peak,
respectively as a ratio of a natural abundance of the isotope to the abundance of the most abundant isotope.
}
