% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{optional_exclude}
\alias{optional_exclude}
\title{Optionaly exlude some entries from data frame (fancier dplyr::anti_join)}
\usage{
optional_exclude(data, exclude, column_name)
}
\arguments{
\item{data}{a data frame}

\item{column_name}{the name of the column where to search for unwanted values}

\item{exlude}{optional vector of values to exclude}
}
\value{
data frame with excluded entries
}
\description{
Optionaly exlude some entries from data frame (fancier dplyr::anti_join)
}
