% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rt_modules.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Plot the clustering and density function.}
\usage{
plot_clustering(pdf, data, peak_indices, plot_lines = FALSE)
}
\arguments{
\item{pdf}{Density estimate using stats::density(...) function. Black line.}

\item{data}{Raw data used to estimate kernel density.}

\item{peak_indices}{Computed indices where to plot the peaks.}

\item{plot_lines}{Whether to plot vertical lines at the cluster centers or not.
Red circles & lines.}
}
\description{
Plot the clustering and density function.
}
