% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_utils.R
\name{reformat_annotation_table}
\alias{reformat_annotation_table}
\title{Reformat annotation table with isotopes from \code{main} to be compatible with \verb{master's} chemical score computing.}
\usage{
reformat_annotation_table(annotation)
}
\arguments{
\item{annotation}{Annotation table after \code{\link[=compute_isotopes]{compute_isotopes()}} with columns
\link[='mz', 'rt', 'rt_cluster', 'module', 'multiple_match', 'expected_mass',
'molecular_formula', 'adduct', 'mass_number_difference', 'monoisotopic_mass',
'mean_intensity', 'mass_defect', 'compound_id']{'mz', 'rt', 'rt_cluster', 'module', 'multiple_match', 'expected_mass', 'molecular_formula', 'adduct', 'mass_number_difference', 'monoisotopic_mass', 'mean_intensity', 'mass_defect', 'compound_id'}}
}
\value{
Reformatted annotation table for use with the \code{master} branch.
}
\description{
Reformat annotation table with isotopes from \code{main} to be compatible with \verb{master's} chemical score computing.
}
