% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_utils.R
\name{reformat_correlation_matrix}
\alias{reformat_correlation_matrix}
\title{Reformat the peak correlation table row and column names to fit with \code{master} branch.}
\usage{
reformat_correlation_matrix(
  peak_table,
  peak_correlation_matrix,
  truncate = FALSE
)
}
\arguments{
\item{peak_table}{Table from which to take the mz & rt values.}

\item{peak_correlation_matrix}{Peak correlation table.}
}
\value{
Peak correlation table where \code{peak} indices are replaced by \code{mz_rt}.
}
\description{
Reformat the peak correlation table row and column names to fit with \code{master} branch.
}
