% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{remove_indipendent_pathways}
\alias{remove_indipendent_pathways}
\title{Remove pathways which are statisticaly independent on significant annotations}
\usage{
remove_indipendent_pathways(pathways, significant)
}
\arguments{
\item{pathways}{data frame of pathway-compound mappings}

\item{significant}{vector of significant compounds}
}
\value{
currated data frame of pathway-compounds mappings
}
\description{
Test whether compounds associated with some specific pathway are equaly
likely to be marked as significant as compounds not assotiated with the
pathway. If \empth{x} is a compound, \emph{P} is some specific pathway and
\emph{S} is a group of significant annotations then the data can be
represented by the following contingency table:
}
\details{
\tabular{lll}{
    \tab x \in P \tab x \notin P \cr
   x \in S \tab a \tab b \cr
   x \notin S \tab c \tab d \cr
}
}
