% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemscore_helpers.R
\name{replace_with_module}
\alias{replace_with_module}
\title{Extract module from Module_RTclust string}
\usage{
replace_with_module(module_rt_clust)
}
\arguments{
\item{module_rt_clust}{Module_RTclust string}
}
\value{
Module portion only
}
\description{
Extract module from Module_RTclust string
}
