% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_parquet}
\alias{save_parquet}
\title{Save data frame to a Parquet file}
\usage{
save_parquet(data, file)
}
\arguments{
\item{data}{Data frame to save.}

\item{file}{Path to the output Parquet file.}
}
\value{
Invisible NULL.
}
\description{
Writes a data frame to disk in Apache Parquet format.
}
