% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelannotationstep4.R
\name{strip_isotope_suffix}
\alias{strip_isotope_suffix}
\title{Strip isotope annotations from adduct names
e.g., "M+H_\link{+1}" -> "M+H"}
\usage{
strip_isotope_suffix(adducts)
}
\description{
Strip isotope annotations from adduct names
e.g., "M+H_\link{+1}" -> "M+H"
}
