% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways.R
\name{update_pathway_scores}
\alias{update_pathway_scores}
\title{Update scores of annotations with pathway scores}
\usage{
update_pathway_scores(annotations, pathways)
}
\arguments{
\item{annotations}{data frame with feature annotations with scores}

\item{pathways}{pathway mappings with computed pathway scores}
}
\value{
augmented annotations
}
\description{
Update scores of annotations with pathway scores
}
