% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_annotation.R
\name{advanced_annotation}
\alias{advanced_annotation}
\title{Wrapper for the advanced annotation steps.}
\usage{
advanced_annotation(
  peak_table,
  compound_table,
  adduct_table = NULL,
  adduct_weights = NULL,
  feature_id_column = NULL,
  intensity_deviation_tolerance = 0.1,
  mass_tolerance = 5e-06,
  isotope_mass_tolerance = NULL,
  mass_defect_tolerance = 0.1,
  mass_defect_precision = 0.01,
  time_tolerance = 10,
  peak_rt_width = 1,
  correlation_threshold = 0.7,
  MplusH_abundance_ratio_check = TRUE,
  multimer_abundance_check = TRUE,
  deep_split = 2,
  min_cluster_size = 10,
  maximum_isotopes = 10,
  min_ions_per_chemical = 2,
  filter_by = c("M-H", "M+H"),
  network_type = "unsigned",
  redundancy_filtering = TRUE,
  pathway_mode = "HMDB",
  pathway_data = NULL,
  excluded_pathways = NULL,
  excluded_pathway_compounds = NULL,
  boosted_compounds = NULL,
  boost_match_by = c("mz", "rt"),
  boost_mass_tolerance = NULL,
  boost_time_tolerance = NULL,
  identify_isotopologues_flag = TRUE,
  level1_primary_adducts = c("M+H", "M-H"),
  outloc = tempdir(),
  n_workers = parallel::detectCores()
)
}
\description{
Wrapper for the advanced annotation steps.
}
