% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_annotation.R
\name{restore_peak_column}
\alias{restore_peak_column}
\title{Restore peak column on annotation after a pipeline stage strips it}
\usage{
restore_peak_column(annotation, peak_restore_map)
}
\arguments{
\item{annotation}{Data frame potentially missing peak column}

\item{peak_restore_map}{Unambiguous (mz, time) -> peak lookup}
}
\value{
annotation with peak column (NA where mz/time was ambiguous)
}
\description{
Restore peak column on annotation after a pipeline stage strips it
}
